requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pandas
streamlit
requests
python-multipart
numpy
matplotlib
scikit-learn